Add files via upload
In this task, you will code the backpropagation algorithm to train a multi-layer perceptron for classifying data into one of four distinct classes. The implementation must be done from scratch, without relying on established libraries such as Tensorflow, Keras, or PyTorch.

Two files, "train data.csv" and "train labels.csv," are provided, containing a dataset of 24,754 samples. Each sample has 784 features, categorized into four classes (0, 1, 2, 3). Your task includes splitting the data into training and validation sets to prevent overfitting. Finally, your model will be evaluated on an unseen test set.

The architecture of your neural network should consist of one input layer, one hidden layer, and one output layer. The labels are in one-hot encoded format, where class 0 is represented as [1, 0, 0, 0], and class 2 is represented as [0, 0, 1, 0]. Ensure the appropriate activation function is used in the output layer. The number of nodes in the hidden layer can be chosen as needed.

You are required to create a single function that enables the utilization of the trained network for predicting the test set. This function should output the labels in a one-hot encoded format, presented as a numpy array.
Neural Net Configuration
1. Hidden Layer
o 784 inputs
o 100 perceptrons
o 100 biases (1 for each perceptrons)
o Activation function - ReLU
2. Output Layer
o 100 inputs (for each neuron from hidden layer)
o 4 perceptrons (to match label data)
o 4 biases
o Activation function – Softmax
For the activation function of the hidden layer we decided to use ReLU for it flexibility and simplicity.
For the output layer we decided to use Softmax to obtain the probability distribution between 0 and 1 to
match with the one-hot label data.